+++ src/multipoles.c
@@ -136,7 +136,7 @@ int coffe_multipoles_init(
             par->sep_len
         );
         mp->l = (int *)coffe_malloc(sizeof(int)*par->multipole_values_len);
-        for (int i = 0; i<par->multipole_values_len; ++i){
+        for (size_t i = 0; i<par->multipole_values_len; ++i){
             mp->l[i] = (int)par->multipole_values[i];
         }
         mp->l_len = (size_t)par->multipole_values_len;
